+++ pubspec.yaml
@@ -44,6 +44,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_driver:
+    sdk: flutter
   flutter_launcher_icons: "^0.7.0"
 
 flutter_icons:
